Modified stepper::tick() to respect DRV8825 timing requirements.
There must be a minimum wait time between changing "DIR" and issuing a pulse at "STEP" of 650ns. Before this change, there was a delay of only 680ns.

I placed the other useful work in the same code block inbetween the two writes, and added an assembly line to force the compiler to respect ordering. The delay is now ~1600ns which leaves some breathing room.

This change made stepper::tick() ~4 clock cycles longer in the worse case. (Redundant reads caused by the barriers)